Extracting service filter from search_form.rb
